requirements.txt
astroid==1.4.9
asttokens==1.1.13
attrs==18.2.0
autologin==0.1.4
Automat==0.7.0
automationhat==0.1.0
beautifulsoup4==4.5.3
bleach==3.0.2
blinker==1.3
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
Cartopy==0.14.2
chardet==2.3.0
click==6.6
colorama==0.3.7
colorzero==1.1
constantly==15.1.0
crochet==1.9.0
cryptography==1.7.1
cssselect==1.0.3
cycler==0.10.0
decorator==4.0.11
defusedxml==0.5.0
devscripts==2.17.6+deb9u2
dj-database-url==0.5.0
Django==2.1.2
django-heroku==0.3.1
docopt==0.6.2
docutils==0.13.1
drumhat==0.1.0
emoji==0.5.1
entrypoints==0.2.3
envirophat==1.0.0
ExplorerHAT==0.4.2
file-magic==0.3.0
Flask==0.10.1
Flask-Admin==1.4.2
flask-heroku==0.1.9
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
formasaurus==0.8.1
fourletterphat==0.1.0
gpiozero==1.5.0
gunicorn==19.9.0
html5lib==0.999999999
hyperlink==18.0.0
idna==2.2
image==1.5.27
incremental==17.5.0
ipykernel==5.1.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==7.4.2
isort==4.2.5
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.8
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
keyring==10.1
keyrings.alt==1.3
Kivy==1.10.1
Kivy-Garden==0.1.4
lazy-object-proxy==1.2.2
lxml==3.7.1
MarkupSafe==0.23
matlab==0.1
matplotlib==2.0.0
mcpi==0.1.1
mistune==0.8.4
mote==0.0.4
motephat==0.0.2
mypy==0.470
nanpy==0.9.6
nbconvert==5.4.0
nbformat==4.4.0
nexmo==2.3.0
notebook==5.7.2
numexpr==2.6.1
numpy==1.12.1
oauthlib==2.0.1
opencv-python==3.4.3.18
paho-mqtt==1.4.0
pandas==0.24.0
pandocfilters==1.4.2
pantilthat==0.0.7
parsel==1.5.1
patsy==0.4.1+dev
pexpect==4.2.1
phatbeat==0.1.1
pianohat==0.1.0
pickleshare==0.7.4
picraft==1.0
piglow==1.2.4
pigpio==1.38
Pillow==4.0.0
prometheus-client==0.5.0
prompt-toolkit==1.0.15
psycopg2==2.7.6.1
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.2.3
pycrypto==2.6.1
PyDispatcher==2.0.5
pyfingerprint==1.5
pygmaps-enhanced==0.2.0
Pygments==2.2.0
pygobject==3.22.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.4.2
pylint==1.6.5
pymatlab==0.2.3
pyOpenSSL==16.2.0
pyparsing==2.1.10
pyperclip==1.5.27
pyserial==3.2.1
pyshp==1.2.10
PySocks==1.6.8
python-apt==1.1.0b5
python-crfsuite==0.9.6
python-dateutil==2.5.3
python-debian==0.1.30
pytz==2016.7
pyxdg==0.25
pyzmq==17.1.2
qtconsole==4.4.3
queuelib==1.5.0
rainbowhat==0.1.0
requests==2.12.4
requests-file==1.4.3
requests-oauthlib==0.7.0
roman==2.0.0
RPi.GPIO==0.6.5
scikit-learn==0.20.2
scipy==0.18.1
Scrapy==1.5.2
scrapy-splash==0.7.2
scrollphat==0.0.7
seaborn==0.7.1
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
Shapely==1.5.17
simplegeneric==0.8.1
simplejson==3.10.0
six==1.12.0
sklearn==0.0
sklearn-crfsuite==0.3.6
skywriter==0.0.7
sn3218==1.2.7
spidev==3.3
SQLAlchemy==1.2.16
tables==3.3.0
tabulate==0.8.3
terminado==0.8.1
testpath==0.4.2
thonny==3.1.0
thonny-pi==1.1
tldextract==2.2.0
tornado==4.5.3
touchphat==0.0.1
tqdm==4.29.1
traitlets==4.3.1
twilio==6.19.1
Twisted==18.9.0
twython==3.4.0
typed-ast==0.6.3
ufw==0.35
unicorn==1.0.1
urllib3==1.19.1
versioneer==0.18
virtualenv==16.1.0
w3lib==1.20.0
Way2sms==1.0.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.11.15
whitenoise==4.1.2
widgetsnbextension==3.4.2
wrapt==1.9.0
WTForms==2.1
zope.interface==4.6.0
